Refactor old bailman
